alici-py/requirements.txt
Flask~=3.1.0
python-dotenv~=1.1.0
openai~=1.77.0
tenacity
tensorflow
numpy~=2.1.3
pillow
psycopg2-binary
flask-sqlalchemy
flask
flask_sqlalchemy
psycopg2
tensorflow
tensorflow-hub
python-dotenv
